Make entry point easily changeable
Clean up the initial starting view code in App delegate into another
tab controller as in code and storyboard so changing starting point
could easy.
