+++ AppDelegate.swift
@@ -16,44 +16,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate,UINavigationControllerDele
 
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
-        // Override point for customization after application launch.
-        window = UIWindow(frame: UIScreen.main.bounds)
         //Connect Firebase
         FIRApp.configure()
-        
-        
-        // Set up the first View Controller
-        let storyboard = UIStoryboard(name: "Louis.Main", bundle: nil)
-        let vc1 = storyboard.instantiateViewController(withIdentifier: "HomeScreen") as! UINavigationController
-        vc1.delegate = self
-        vc1.tabBarItem.title = "Add Word"
-        vc1.tabBarItem = UITabBarItem.init(tabBarSystemItem: UITabBarSystemItem.bookmarks, tag: 0)
-        
-        // Set up the second View Controller
-        let vc2 = storyboard.instantiateViewController(withIdentifier: "HomeScreen") as! UINavigationController
-        vc2.delegate = self
-        vc2.tabBarItem.title = "Word Set"
-        vc2.tabBarItem = UITabBarItem.init(tabBarSystemItem: UITabBarSystemItem.history, tag: 1)
-        
-        let vc3 = storyboard.instantiateViewController(withIdentifier: "HomeScreen") as! UINavigationController
-        vc3.delegate = self
-        vc3.tabBarItem.title = "Media"
-        vc3.tabBarItem = UITabBarItem.init(tabBarSystemItem: UITabBarSystemItem.favorites, tag: 2)
-        
-        let vc4 = storyboard.instantiateViewController(withIdentifier: "HomeScreen") as! UINavigationController
-        vc4.delegate = self
-        vc4.tabBarItem.title = "Personal"
-        vc4.tabBarItem = UITabBarItem.init(tabBarSystemItem: UITabBarSystemItem.more, tag: 3)
-        
-        
-        
-        // Set up the Tab Bar Controller to have four tabs
-        let tabBarController = UITabBarController()
-        tabBarController.viewControllers = [vc1, vc2, vc3, vc4]
-        
-        // Make the Tab Bar Controller the root view controller
-        window?.rootViewController = tabBarController
-        window?.makeKeyAndVisible()
         return true
     }
